Create README.md
Wanderlust is a modern and user-friendly vacation rental platform designed to simplify the process of booking and hosting accommodations worldwide. Whether you're a traveller looking for a cosy stay or a host aiming to rent out your property, Wanderlust provides a seamless experience with an intuitive UI and secure transactions.